pyproject.toml
[project]
name = "protein-property-vectors"
version = "0.1.0"
description = "Task arithmetic for protein fitness prediction via LoRA fine-tuning on ESM-2"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "wandb>=0.16.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "biopython>=1.82",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "ipykernel>=6.27.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
